requirements.txt
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
docutils==0.18.1
Faker==9.8.3
idna==3.3
importlib-metadata==4.8.2
keyring==23.4.0
mongoengine==0.23.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pkginfo==1.8.1
protobuf==3.19.1
psycopg2-binary==2.9.2
pydantic==1.8.2
Pygments==2.10.0
pymongo==3.12.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
readme-renderer==30.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
text-unidecode==1.3
tqdm==4.62.3
twine==3.6.0
typing-extensions==4.0.1
urllib3==1.26.7
webencodings==0.5.1
zipp==3.6.0
